+++ Models.py
@@ -1,0 +1,59 @@
+import torch
+import torch.nn as nn
+from torch.distributions import Categorical
+from stable_baselines3.common.preprocessing import preprocess_obs
+import numpy as np
+
+class LogUNet(nn.Module):
+    def __init__(self, env, device='cuda', hidden_dim=256):
+        super(LogUNet, self).__init__()
+        self.env = env
+        self.device = device
+        try:
+            self.nS = env.observation_space.n
+        except AttributeError:
+            self.nS = env.observation_space.shape[0]
+
+        self.nA = env.action_space.n
+        self.fc1 = nn.Linear(self.nS, hidden_dim, device=self.device)
+        self.fc2 = nn.Linear(hidden_dim, hidden_dim, device=self.device)
+        self.fc3 = nn.Linear(hidden_dim, self.nA, device=self.device)
+        self.relu = nn.ReLU()
+        # self.relu = nn.LeakyReLU()
+     
+    def forward(self, x):
+        if not isinstance(x, torch.Tensor):
+            x = torch.tensor(x, device=self.device, dtype=torch.float32).detach()  # Convert to PyTorch tensor
+
+        x = preprocess_obs(x, self.env.observation_space)
+
+        x = self.fc1(x)
+        x = self.relu(x)
+        x = self.fc2(x)
+        x = self.relu(x)
+        x = self.fc3(x)
+
+        return x
+
+    def get_chi(self, logu_a):
+        prior_policy = 1 / self.nA
+        chi = torch.sum(prior_policy * torch.exp(logu_a))
+        return chi
+        
+    def choose_action(self, state, greedy=False):
+        with torch.no_grad():
+            logu = self.forward(state)
+
+            if greedy:
+                a = logu.argmax()
+                return a.item()
+
+            # First subtract a baseline:
+            logu = logu - (torch.max(logu) + torch.min(logu))/2
+            dist = torch.exp(logu) * 1 / self.nA
+            dist = dist / torch.sum(dist)
+            c = Categorical(dist)
+            a = c.sample()
+
+        return a.item()
+
